review CRUD complete with validation

+++ index.js
@@ -35,20 +35,98 @@ chillGamer = async () => {
         // REVIEWS RELATED API"S                           //
         // --------------------------------------------------
         app.get("/reviews", async (req, res) => {
-            const data = await reviews.find({}).toArray();
-            res.send(data);
+            const result = await reviews.find({}).toArray();
+            if (result.length === 0) {
+                return res.status(404).json({ error: "No Data Found" });
+            }
+            res.status(200).json({
+                message: `Reviews Found = ${result.length}`,
+                result,
+            });
         });
 
+        // Add review
         app.post("/reviews", async (req, res) => {
             const data = req.body;
-            const result = await reviews.insertOne(data);
-            res.send(result);
+            const addReviews = {};
+
+            // Validate Data
+            for (const [key, value] of Object.entries(data)) {
+                if (typeof value === "string" && value.trim() !== "") {
+                    addReviews[key] = value.trim();
+                } else if (typeof value === "number" && !isNaN(value)) {
+                    addReviews[key] === value;
+                }
+            }
+
+            // Check if addReviews is Empty
+            if (Object.keys(addReviews).length === 0) {
+                return res.status(400).json({ error: "No valid field found" });
+            }
+
+            const result = await reviews.insertOne(addReviews);
+            res.status(200).json({
+                message: "Reviews added successfully",
+                result,
+            });
         });
 
+        // Update review
         app.put("/reviews/:id", async (req, res) => {
             const id = req.params.id;
+            const data = req.body;
             const filter = { _id: new ObjectId(id) };
-            const {} = req.body;
+            let updateReviews = {};
+
+            // data validation
+            // Check whitch field is true
+            for (const [key, value] of Object.entries(data)) {
+                if (typeof value === "string" && value.trim() !== "") {
+                    updateReviews[key] = value.trim();
+                } else if (typeof value === "number" && !isNaN(value)) {
+                    updateReviews[key] = value;
+                }
+            }
+
+            // Check if updateData is empty
+            if (Object.keys(updateReviews).length === 0) {
+                return res.status(400).json({ error: "No Valid Field Found" });
+            }
+
+            const result = await reviews.updateOne(
+                filter,
+                {
+                    $set: updateReviews,
+                },
+                { upsert: true }
+            );
+            res.status(200).json({
+                message: "Review Updated Successfully",
+                result,
+            });
+        });
+
+        //  Delete review
+        app.delete("/reviews/:id", async (req, res) => {
+            const id = req.params.id;
+            if (!ObjectId.isValid(id)) {
+                return res
+                    .status(400)
+                    .json({ error: "Invalid Review ID Format" });
+            }
+
+            const query = { _id: new ObjectId(id) };
+            const reviewExist = await reviews.findOne(query);
+
+            if (!reviewExist) {
+                return res.status(404).json({ errpr: "Review Not Found" });
+            }
+
+            const result = await reviews.deleteOne(query);
+            res.status(200).json({
+                message: "Review Deleted Successfully",
+                result,
+            });
         });
 
         // --------------------------------------------------
@@ -56,7 +134,13 @@ chillGamer = async () => {
         // --------------------------------------------------
         app.get("/users", async (req, res) => {
             const result = await users.find({}).toArray();
-            res.send(result);
+            if (result.length === 0) {
+                return res.status(404).json({ error: "No Data Found" });
+            }
+            res.status(200).json({
+                message: `Users Found = ${result.length}`,
+                result,
+            });
         });
 
         //  Create a user
@@ -78,12 +162,15 @@ chillGamer = async () => {
             // Check if the addUser is empty
             if (Object.keys(addUser).length === 0) {
                 return res
-                    .status(404)
-                    .json({ error: "No Valid User Data Found" });
+                    .status(400)
+                    .json({ error: "No Valid Fields Provided" });
             }
 
             const result = await users.insertOne(addUser);
-            res.send(result);
+            res.status(200).json({
+                message: "User Added Successfully",
+                result,
+            });
         });
 
         // Update users Info
@@ -105,21 +192,43 @@ chillGamer = async () => {
 
             // Check if updateData is empty
             if (Object.keys(updateData).length === 0) {
-                return res.status(404).json({ error: "No Valid Data Found" });
+                return res.status(400).json({ error: "No Valid Field Found" });
             }
 
-            const result = await users.updateOne(filter, {
-                $set: updateData,
+            const result = await users.updateOne(
+                filter,
+                {
+                    $set: updateData,
+                },
+                { upsert: true }
+            );
+            res.status(200).json({
+                message: "User Updated Successfully",
+                result,
             });
-            res.send(result);
         });
 
         // Delete an User
         app.delete("/users/:id", async (req, res) => {
             const id = req.params.id;
+            if (!ObjectId.isValid(id)) {
+                return res
+                    .status(400)
+                    .json({ error: "Invalid User ID Format" });
+            }
+
             const query = { _id: new ObjectId(id) };
+            const userExist = await users.findOne(query);
+
+            if (!userExist) {
+                return res.status(404).json({ errpr: "User Not Found" });
+            }
+
             const result = await users.deleteOne(query);
-            res.send(result);
+            res.status(200).json({
+                message: "User Deleted Successfully",
+                result,
+            });
         });
     } catch {
         console.dir;
@@ -128,7 +237,10 @@ chillGamer = async () => {
 chillGamer();
 
 app.get("/", (req, res) => {
-    res.send("Welcome to Chill Gamer. A Game Review Application");
+    res.status(200).json({
+        message: "Welcome to Chill Gamer. A Game Review Application API's",
+        Url: "https://chill-gamer-7f9f1.web.app",
+    });
 });
 app.listen(port, () => {
     console.log("Chill Gamer listening: ", port);
